+++ package/lib/src/ui/models/input_field.dart
@@ -94,6 +94,12 @@ class InputField {
   /// this is a `DateTime` that represents a maximum date allowed to be captured by a `DATE` field
   DateTime? maxDate;
 
+  /// this is selected levels field
+  List<int> allowedSelectedLevels;
+
+  /// State variable for auto updating location
+  bool? disableLocationAutoUpdate;
+
   ///
   /// This is the default constructor to the `InputField`
   ///
@@ -126,6 +132,8 @@ class InputField {
     this.hasError = false,
     this.allowFuturePeriod = false,
     this.disablePastPeriod = false,
+    this.allowedSelectedLevels = const [],
+    this.disableLocationAutoUpdate = false,
   });
 
   /// `InputField.toString()` is the method to return the `String` representation of the `InputField`